split app into multiple components
